Fixing merge
