+++ earn/src/components/portfolio/PortfolioMetrics.tsx
@@ -1,9 +1,13 @@
+import { useState } from 'react';
+
 import { Display, Text } from 'shared/lib/components/common/Typography';
 
 import { TokenData } from '../../data/TokenData';
 import { TokenBalance } from '../../pages/PortfolioPage';
+import { rgba } from '../../util/Colors';
 import { formatTokenAmount, roundPercentage } from '../../util/Numbers';
 import { APYContainer, BalanceContainer, PieChartContainer } from './PortfolioGrid';
+import PortfolioPieChartWidget, { PortfolioPieChartSlice } from './PortfolioPieChartWidget';
 
 export type PortfolioMetricsProps = {
   balances: TokenBalance[];
@@ -16,18 +20,43 @@ export type PortfolioMetricsProps = {
  * This includes a pie, balance, and APY.
  */
 export default function PortfolioMetrics(props: PortfolioMetricsProps) {
-  const { balances, activeAsset } = props;
+  const { balances, activeAsset, activeColor } = props;
+
+  const [activeIndex, setActiveIndex] = useState<number>(-1);
 
   const activeBalances = balances.filter(
     (balance) => activeAsset && balance.token.referenceAddress === activeAsset.referenceAddress
   );
   const totalTokenBalance = activeBalances.reduce((acc, balance) => acc + balance.balance, 0);
+  const activeSlices: PortfolioPieChartSlice[] = activeBalances.map((balance: TokenBalance, i: number) => {
+    return {
+      token: balance.token,
+      percent: balance.balance / totalTokenBalance || 0,
+      color: activeColor ? rgba(activeColor, (i + 1) / activeBalances.length) : 'transparent',
+      isKitty: balance.isKitty,
+      pairName: balance.pairName,
+      index: i,
+    };
+  });
+  const activeSlice = activeIndex >= 0 ? activeSlices[activeIndex] : null;
   let totalBalance = totalTokenBalance;
   let apySum = activeBalances.reduce((acc, balance) => acc + balance.apy * balance.balance, 0);
   let apy = apySum / totalBalance || 0;
+  if (activeSlice) {
+    totalBalance = activeBalances[activeIndex].balance;
+    apySum = activeBalances[activeIndex].apy;
+    apy = apySum;
+  }
   return (
     <>
-      <PieChartContainer>{/* TODO */}</PieChartContainer>
+      <PieChartContainer>
+        <PortfolioPieChartWidget
+          activeIndex={activeIndex}
+          setActiveIndex={setActiveIndex}
+          slices={activeSlices}
+          token={activeAsset}
+        />
+      </PieChartContainer>
       <BalanceContainer>
         <Text size='M' weight='bold' color='rgba(130, 160, 182, 1)'>
           Balance
